+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ---
 
 - ✨ Add Service Worker
+
+2018-06-10
+
+---
+
+- ✨ Add Apollo & Graph QL
+- ✨ Add Friendly error message